Allow `JavaTemplate` to use `JavaTypeCache`
As currently every application of `JavaTemplate` leads to a new `JavaParser` getting created with its own `JavaTypeCache` the result is that applying the same `JavaTemplate` multiple times to a single source file will potentially end up type-attributing the LST with multiple identical copies of `JavaType` objects, which in the end causes a lot of memory overhead.

This commit aims at mitigating this issue by allowing `JavaTemplate` (specifically `JavaTemplateParser`) to use a `JavaTypeCache`. The life cycle of the `JavaTypeCache` is the same as that of the `JavaTemplate` and thus no longer than that of the recipe and will thus also not allow reuse across source files. This could potentially be addressed by adding a `JavaTemplate.Builder#context(ExecutionContext)` method and then extend `ExecutionContext` to allow caching objects which are keyed on arbitrary objects rather than just `String`.